requirements.txt
# Fraud Detection Claims - Root Requirements
# For Databricks notebooks and jobs

# Databricks SDK and CLI
databricks-sdk>=0.18.0
databricks-cli>=0.18.0

# Data Processing
pyspark>=3.4.0
pandas>=2.0.0
numpy>=1.24.0

# LangChain and LangGraph
langgraph>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
databricks-langchain
unitycatalog-langchain[databricks]

# Vector Search
databricks-vectorsearch

# ML and AI
mlflow>=2.9.0

# Utilities
pyyaml>=6.0
backoff>=2.2.0
python-dotenv>=1.0.0


